% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transformations.R
\name{BestTransform}
\alias{BestTransform}
\title{Transforms the dataset using the best possible transformation technique}
\usage{
BestTransform(data, dv)
}
\arguments{
\item{data}{Any data frame that has atlest one column with continuous data and that has to transformed into normal form}

\item{dv}{Dependent variable in the dataset.}
}
\value{
Returns a list of 2 objects:
\describe{
\item{trans_data}{Tranformed Dataset for all Continuous variables}
\item{trans_fit_model}{Model fit file to fit the test data}
}
}
\description{
Considers the best transformed dataset based on three metrics
\itemize{
\item Shapiro P Value
\item Pearson P Value
\item Min skewness
}
Best possible transformation is applied to all the continous columns in the dataset for each of the above mentioned metrics and three datasets are collected respectively. Best dataset is chosen based on the performance metrics of the dataset on a base model. Original dataset is also considered for the performance measure to actually check if transforming the data is necessary.
}
\details{
Takes in a data frame and performs the best possible transformation to each of the columns in the data frame
}
