% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformationPrerequisites.R
\name{model_my_data}
\alias{model_my_data}
\title{Perform Linear model on the provided dataset}
\usage{
model_my_data(data, num)
}
\arguments{
\item{data}{A list of datasets}

\item{num}{Index of the dataset to be accessed}
}
\value{
The performance measures of the dataset
}
\description{
Takes in a dataset gives the perfoormance measures by taking a 75:25 train-test split. Performance measures are considered based the distribution of the DV column
}
